requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.10
numpy==2.0.2
packaging==24.1
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
thinc==8.3.2
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
